requirements.txt
requests~=2.25.1
djangorestframework~=3.6.2
selenium~=3.141.0
